+++ lib/src/exceptions/no_such_element_exception.dart
@@ -1,1 +1,4 @@
+/// Thrown when an [Optional] value is accessed but no data is present.
+///
+/// See [Optional.data] and [Optional.orElseThrow].
 class NoSuchElementException implements Exception {}